+++ ClientProject/Mof/RachetNClankCopy/RachetApplication/Camera/CameraController.cpp
@@ -2,7 +2,6 @@
 
 #include "My/Core/Define.h"
 #include "Camera.h"
-#include "../Character.h"
 
 
 void my::CameraController::UpdateCameraPosition(const std::shared_ptr<my::Camera>& camera) {
