+++ ClientProject/Mof/RachetNClankCopy/RachetApplication/Character/Enemy.h
@@ -1,0 +1,13 @@
+#ifndef MY_ENEMY_H
+#define MY_ENEMY_H
+
+
+#include "../Actor.h"
+
+
+namespace my {
+class Enemy : public my::Actor {
+    using super = my::Actor;
+};
+}
+#endif // !MY_ENEMY_H
