--- ClientProject/Mof/RachetNClankCopy/RachetApplication/QuickChangeMenu.h
@@ -1,55 +1,0 @@
-#ifndef MY_QUICK_CHANGE_MENU_H
-#define MY_QUICK_CHANGE_MENU_H
-
-
-#include "My/UI/UIPanel.h"
-
-#include "My/Core/Observer.h"
-#include "QuickChangeSystem.h"
-
-
-namespace my {
-class QuickChangeMenuItem : public my::UIItem /*, public my::Observer<my::QuickChangeSystemItem> */{
-    using super = my::UIItem;
-public:
-    /// <summary>
-    /// コンストラクタ
-    /// </summary>
-    /// <param name="name"></param>
-    QuickChangeMenuItem(const char* name);
-    /// <summary>
-    /// 通知イベント
-    /// </summary>
-    /// <param name="observable"></param>
-    /// <param name="event"></param>
-    //virtual void OnNotify(const std::shared_ptr<my::QuickChangeSystem>& observable, const char* event) override;
-    /// <summary>
-    /// 描画
-    /// </summary>
-    /// <param name=""></param>
-    /// <returns></returns>
-    virtual bool Render(void) override;
-};
-class QuickChangeMenu : public my::UIPanel , public my::Observer<my::QuickChangeSystem> {
-    using super = my::UIPanel;
-public:
-    /// <summary>
-    /// コンストラクタ
-    /// </summary>
-    /// <param name="name"></param>
-    QuickChangeMenu(const char* name);
-    /// <summary>
-    /// 通知イベント
-    /// </summary>
-    /// <param name="observable"></param>
-    /// <param name="event"></param>
-    virtual void OnNotify(const std::shared_ptr<my::QuickChangeSystem>& observable, const char* event) override;
-    /// <summary>
-    /// 描画
-    /// </summary>
-    /// <param name=""></param>
-    /// <returns></returns>
-    virtual bool Render(void) override;
-};
-}
-#endif // !MY_QUICK_CHANGE_MENU_H
