+++ ClientProject/Mof/RachetNClankCopy/RachetApplication/RachetApplication.vcxproj
@@ -147,6 +147,7 @@
   <ItemGroup>
     <ClCompile Include="..\include\My\Core\State.cpp" />
     <ClCompile Include="..\include\My\Core\StateMachine.cpp" />
+    <ClCompile Include="..\include\My\UI\CanvasLocator.cpp" />
     <ClCompile Include="..\include\My\UI\UICanvas.cpp" />
     <ClCompile Include="..\include\My\UI\UIItem.cpp" />
     <ClCompile Include="..\include\My\UI\UIPanel.cpp" />
@@ -155,21 +156,21 @@
     <ClCompile Include="Camera\CameraController.cpp" />
     <ClCompile Include="Camera\CameraLocator.cpp" />
     <ClCompile Include="Camera\CameraManager.cpp" />
-    <ClCompile Include="CanvasLocator.cpp" />
-    <ClCompile Include="Character.cpp" />
+    <ClCompile Include="Character\Character.cpp" />
+    <ClCompile Include="Character\Enemy.cpp" />
+    <ClCompile Include="Character\Player.cpp" />
     <ClCompile Include="GameApp.cpp" />
-    <ClCompile Include="GameMoney.cpp" />
     <ClCompile Include="Gamepad.cpp" />
+    <ClCompile Include="GameSystem\GameMoney.cpp" />
+    <ClCompile Include="GameSystem\QuickChangeSystem.cpp" />
+    <ClCompile Include="GameSystem\Save\SaveData.cpp" />
+    <ClCompile Include="GameSystem\Save\SaveSystem.cpp" />
+    <ClCompile Include="GameSystem\WeaponSystem.cpp" />
     <ClCompile Include="Game\GameManager.cpp" />
     <ClCompile Include="Game\Renderer.cpp" />
-    <ClCompile Include="Player.cpp" />
-    <ClCompile Include="QuickChangeMenu.cpp" />
-    <ClCompile Include="QuickChangeSystem.cpp" />
     <ClCompile Include="ResourceLocator.cpp" />
-    <ClCompile Include="Save\SaveData.cpp" />
-    <ClCompile Include="Save\SaveSystem.cpp" />
-    <ClCompile Include="Stage.cpp" />
-    <ClCompile Include="WeaponSystem.cpp" />
+    <ClCompile Include="Stage\Stage.cpp" />
+    <ClCompile Include="UI\QuickChangeMenu.cpp" />
     <ClCompile Include="Weapon\BombGlove.cpp" />
     <ClCompile Include="Weapon\Mechanical.cpp" />
     <ClCompile Include="Weapon\OmniWrench.cpp" />
@@ -178,6 +179,7 @@
     <ClCompile Include="WinMain.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="..\include\My\Core\AbstructContainer.h" />
     <ClInclude Include="..\include\My\Core\Define.h" />
     <ClInclude Include="..\include\My\Core\Header.h" />
     <ClInclude Include="..\include\My\Core\Math.h" />
@@ -188,35 +190,35 @@
     <ClInclude Include="..\include\My\Core\StateMachine.h" />
     <ClInclude Include="..\include\My\Core\Trait.h" />
     <ClInclude Include="..\include\My\Core\Utility.h" />
+    <ClInclude Include="..\include\My\UI\CanvasLocator.h" />
     <ClInclude Include="..\include\My\UI\UICanvas.h" />
     <ClInclude Include="..\include\My\UI\UIItem.h" />
     <ClInclude Include="..\include\My\UI\UIPanel.h" />
-    <ClInclude Include="AbstructContainer.h" />
     <ClInclude Include="Actor.h" />
     <ClInclude Include="Camera\Camera.h" />
     <ClInclude Include="Camera\CameraController.h" />
     <ClInclude Include="Camera\CameraLocator.h" />
     <ClInclude Include="Camera\CameraManager.h" />
-    <ClInclude Include="CanvasLocator.h" />
-    <ClInclude Include="Character.h" />
+    <ClInclude Include="Character\Character.h" />
+    <ClInclude Include="Character\Enemy.h" />
+    <ClInclude Include="Character\Player.h" />
     <ClInclude Include="Factory\Creator.h" />
     <ClInclude Include="Factory\Factory.h" />
     <ClInclude Include="GameDefine.h" />
     <ClInclude Include="GameApp.h" />
-    <ClInclude Include="GameMoney.h" />
     <ClInclude Include="Gamepad.h" />
+    <ClInclude Include="GameSystem\GameMoney.h" />
+    <ClInclude Include="GameSystem\QuickChangeSystem.h" />
+    <ClInclude Include="GameSystem\Save\SaveData.h" />
+    <ClInclude Include="GameSystem\Save\SaveSystem.h" />
+    <ClInclude Include="GameSystem\WeaponSystem.h" />
     <ClInclude Include="Game\GameManager.h" />
     <ClInclude Include="Game\Renderer.h" />
-    <ClInclude Include="Player.h" />
-    <ClInclude Include="QuickChangeMenu.h" />
-    <ClInclude Include="QuickChangeSystem.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="ResourceLocator.h" />
     <ClInclude Include="ResourceManager.h" />
-    <ClInclude Include="Save\SaveData.h" />
-    <ClInclude Include="Save\SaveSystem.h" />
-    <ClInclude Include="Stage.h" />
-    <ClInclude Include="WeaponSystem.h" />
+    <ClInclude Include="Stage\Stage.h" />
+    <ClInclude Include="UI\QuickChangeMenu.h" />
     <ClInclude Include="Weapon\BombGlove.h" />
     <ClInclude Include="Weapon\Mechanical.h" />
     <ClInclude Include="Weapon\OmniWrench.h" />
